+++ app/src/main/java/com/maxfour/music/loader/PlaylistSongLoader.java
@@ -15,7 +15,7 @@ import java.util.List;
 public class PlaylistSongLoader {
 
     @NonNull
-    public static List<PlaylistSong> getPlaylistSongList(@NonNull final Context context, final int playlistId) {
+    public static List<PlaylistSong> getPlaylistSongList(@NonNull final Context context, final long playlistId) {
         List<PlaylistSong> songs = new ArrayList<>();
         Cursor cursor = makePlaylistSongCursor(context, playlistId);
 
@@ -31,24 +31,24 @@ public class PlaylistSongLoader {
     }
 
     @NonNull
-    private static PlaylistSong getPlaylistSongFromCursorImpl(@NonNull Cursor cursor, int playlistId) {
-        final int id = cursor.getInt(0);
+    private static PlaylistSong getPlaylistSongFromCursorImpl(@NonNull Cursor cursor, long playlistId) {
+        final long id = cursor.getLong(0);
         final String title = cursor.getString(1);
         final int songNumber = cursor.getInt(2);
         final int year = cursor.getInt(3);
         final long duration = cursor.getLong(4);
         final String data = cursor.getString(5);
         final int dateModified = cursor.getInt(6);
-        final int albumId = cursor.getInt(7);
+        final long albumId = cursor.getLong(7);
         final String albumName = cursor.getString(8);
-        final int artistId = cursor.getInt(9);
+        final long artistId = cursor.getLong(9);
         final String artistName = cursor.getString(10);
-        final int idInPlaylist = cursor.getInt(11);
+        final long idInPlaylist = cursor.getLong(11);
 
         return new PlaylistSong(id, title, songNumber, year, duration, data, dateModified, albumId, albumName, artistId, artistName, playlistId, idInPlaylist);
     }
 
-    public static Cursor makePlaylistSongCursor(@NonNull final Context context, final int playlistId) {
+    public static Cursor makePlaylistSongCursor(@NonNull final Context context, final long playlistId) {
         try {
             return context.getContentResolver().query(
                     MediaStore.Audio.Playlists.Members.getContentUri("external", playlistId),
