+++ app/src/main/java/com/maxfour/music/util/MusicUtil.java
@@ -42,13 +42,13 @@ import java.util.regex.Pattern;
 
 public class MusicUtil {
 
-    public static Uri getMediaStoreAlbumCoverUri(int albumId) {
+    public static Uri getMediaStoreAlbumCoverUri(long albumId) {
         final Uri sArtworkUri = Uri.parse("content://media/external/audio/albumart");
 
         return ContentUris.withAppendedId(sArtworkUri, albumId);
     }
 
-    public static Uri getSongFileUri(int songId) {
+    public static Uri getSongFileUri(long songId) {
         return ContentUris.withAppendedId(MediaStore.Audio.Media.EXTERNAL_CONTENT_URI, songId);
     }
 
@@ -178,7 +178,7 @@ public class MusicUtil {
         return songNumberToFix % 1000;
     }
 
-    public static void insertAlbumArt(@NonNull Context context, int albumId, String path) {
+    public static void insertAlbumArt(@NonNull Context context, long albumId, String path) {
         ContentResolver contentResolver = context.getContentResolver();
 
         Uri artworkUri = Uri.parse("content://media/external/audio/albumart");
@@ -192,7 +192,7 @@ public class MusicUtil {
         contentResolver.notifyChange(artworkUri, null);
     }
 
-    public static void deleteAlbumArt(@NonNull Context context, int albumId) {
+    public static void deleteAlbumArt(@NonNull Context context, long albumId) {
         ContentResolver contentResolver = context.getContentResolver();
         Uri localUri = Uri.parse("content://media/external/audio/albumart");
         contentResolver.delete(ContentUris.withAppendedId(localUri, albumId), null, null);
@@ -244,7 +244,7 @@ public class MusicUtil {
                 // as from the album art cache
                 cursor.moveToFirst();
                 while (!cursor.isAfterLast()) {
-                    final int id = cursor.getInt(0);
+                    final long id = cursor.getLong(0);
                     final Song song = SongLoader.getSong(context, id);
                     MusicPlayerRemote.removeFromQueue(song);
                     cursor.moveToNext();
@@ -253,7 +253,7 @@ public class MusicUtil {
                 // Step 2: Remove files from card
                 cursor.moveToFirst();
                 while (!cursor.isAfterLast()) {
-                    final int id = cursor.getInt(0);
+                    final long id = cursor.getLong(0);
                     final String name = cursor.getString(1);
                     try { // File.delete can throw a security exception
                         final File f = new File(name);
